Crate test_parachain_undying

Basic parachain that adds a number as part of its state.

Structs

BlockData

Block data for this parachain.

GraveyardState

Block data for this parachain.

HeadData

Head data for this parachain.

StateMismatch

Start state mismatched with parent header’s state hash.

Constants

WASM_BINARY

WASM_BINARY_BLOATY

Functions

execute

Execute a block body on top of given parent head, producing new parent head and new state if valid.

execute_transaction

Executes all graveyard transactions in the block.

hash_state

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.